Loading Model to screen
Loading a GLTF model from mixamo to screen. You cannot save models underl a folder called models in public for some reason it does not recognize it and we have to fix this issue. When moving the textures folder you have to update it in the gltf file. Created a JSX component for the model to be able to render the skelton, texture, and skin properly